Create dedicated ApparatusOption component
