+++ HarpGuru/OverlayMenu/MenuOptions/ApparatusOption/index.ts
@@ -1,0 +1,1 @@
+export { ApparatusOption } from './components'
