+++ HarpGuru/OverlayMenu/MenuOptions/index.ts
@@ -1,3 +1,4 @@
+export { ApparatusOption } from './ApparatusOption'
 export { HarpKeyOption } from './HarpKeyOption'
 export { PozitionOption } from './PozitionOption'
 export { RootPitchOption } from './RootPitchOption'
